pyproject.toml
[project]
name = "map-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.15.0",
    "faiss-cpu>=1.11.0",
    "google-genai>=1.19.0",
    "google-generativeai>=0.8.5",
    "image>=1.5.33",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
    "tabulate>=0.9.0",
]

[tool.uv.workspace]
members = [
    "app",
]
